Improve features.h, improve errno logging and fix ssize_t comparisons
- Change all function declarations with internal visibility
  from HIDDEN to HIDDEN_A using the new macro in features.h,
  ensuring consistent symbol hiding and easier attribute logic.
- All syslog/logging calls using %m now print strerror(errno)
  for accurate error messages.
- Use ssize_t (with explicit casts where needed)
  for snprintf/array size, loop and conditions handling,
  preventing signed/unsigned comparison warnings
  and potential overflows.
- Extend/fix resource and netgroup conditional logic for safer builds
  when BSDDB or YP are used.
- Minor code cleanups: correct for possible negative lengths,
  silence warnings in malloc/strncpy, initialize local variables,
  and clarify type correctness in utility and portable headers.
- No external API or ABI change. Purely visibility, portability,
  and robust error diagnostics improvements.

+++ cryptutil.c
@@ -70,7 +70,7 @@ crypt_newhash(const char *pass, const char *pref, char *hash, size_t hashlen)
 	if (pref == NULL)
 		pref = defaultpref;
 
-	for (i = 0; i < maxchoice; i++) {
+	for (i = 0; i < (ssize_t)maxchoice; i++) {
 		const char *choice = choices[i];
 		size_t len = strlen(choice);
 		if (strcmp(pref, choice) == 0) {
@@ -90,7 +90,7 @@ crypt_newhash(const char *pass, const char *pref, char *hash, size_t hashlen)
 			break;
 		}
 	}
-	if (i == maxchoice) {
+	if (i == (ssize_t)maxchoice) {
 		errno = EINVAL;
 		goto err;
 	}
